dlm-yarg/binary: int size warning

+++ cyarg/main.c
@@ -80,7 +80,7 @@ int main(int argc, const char* argv[]) {
     vmHost.argv = argv;
     vmHost.exitCode = EX_OK;
     if (libPath) {
-        vm.libraryPath = copyString(libPath, strlen(libPath));
+        vm.libraryPath = copyString(libPath, (int)strlen(libPath));
     }
     
     int returnCode = EX_OK;
